Example Of Logged Data Drawn From Adafuit IO
Shows example MQQT data logging to external server.
The data represents a number of days. In this example, overnight the unit is arranged to be powered from its internal battery. For voltage drops of less than 0.1 volt for the 2000 mAh battery, the heating effect from re-charging can be seen to be minimal. For longer periods i.e. larger volt drops, a temporary impact can be identified in the measured temperature data. The data capture aligns with the thermal measurments, confirming the optimal period of un-powered use where a battery for backup purposes might be employed.